refactor: add hook for layout, env, and remove routes to separate component
