+++ src/App.js
@@ -1,17 +1,12 @@
 import React from "react";
 import Layout from "./components/Layout";
-import {Route, Routes} from "react-router-dom";
-import Main from "./components/Main";
-import Widgets from "./components/Widgets";
+import RoutesComponent from "./routes/RoutesComponent";
 
 function App() {
     return (
         <div>
             <Layout/>
-            <Routes>
-                <Route path='main' element={<Main/>}/>
-                <Route path='widgets' element={<Widgets/>}/>
-            </Routes>
+            <RoutesComponent />
         </div>
     );
 }
